Get active projr profile

Get active projr profile. For each setting, preference is given if specified in _projr-local.yml and, failing that, in _projr-<profile>.yml (or in the build-<profile> and directories-<profile> keys). Note that the active projr profile is only equal to the PROJR_PROFILE environment variable if the latter is set and the corresponding profile exists (either as keys in _projr.yml or as a file _projr-<profile>.yml).

Usage

projr_profile_get()

Value

Character vector of length 1 corresponding to active projr profile. If "default" is returned, then no profile is active. Note that any _projr-local.yml file will always overwrite profile-specific and default settings.